+++ Runtime/Scripts/Interaction/Locomotion/RedirectedWalking.cs
@@ -18,7 +18,7 @@ namespace SentienceLab
 	/// </summary>
 	/// 
 
-	[AddComponentMenu("SentienceLab/Interaction/XR/Redirected Walking")]
+	[AddComponentMenu("SentienceLab/Interaction/Locomotion/Redirected Walking")]
 	public class RedirectedWalking : MonoBehaviour
 	{
 		[Tooltip("The node that represents the user's head")]