+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Moved locomotion-related scripts into different menu folder
+- Reworked Teleportation related scripts (only one controller script, separate renderer)
+
 ### Removed
 
 - Moving MajorDomo into own Unity package